Move validation function to common.go
