+++ tests/cel/clientsettingspolicy_test.go
@@ -1,17 +1,14 @@
 package cel
 
 import (
-	"context"
 	"testing"
 
 	. "github.com/onsi/gomega"
 	controllerruntime "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 	gatewayv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 
 	ngfAPIv1alpha1 "github.com/nginx/nginx-gateway-fabric/v2/apis/v1alpha1"
 	"github.com/nginx/nginx-gateway-fabric/v2/internal/framework/helpers"
-	"github.com/nginx/nginx-gateway-fabric/v2/tests/framework"
 )
 
 func TestClientSettingsPoliciesTargetRefKind(t *testing.T) {
@@ -78,7 +75,16 @@ func TestClientSettingsPoliciesTargetRefKind(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			validateClientSettingsPolicy(t, tt, g, k8sClient)
+			policySpec := tt.policySpec
+			policySpec.TargetRef.Name = gatewayv1alpha2.ObjectName(uniqueResourceName(testTargetRefName))
+			clientSettingsPolicy := &ngfAPIv1alpha1.ClientSettingsPolicy{
+				ObjectMeta: controllerruntime.ObjectMeta{
+					Name:      uniqueResourceName(testPolicyName),
+					Namespace: defaultNamespace,
+				},
+				Spec: policySpec,
+			}
+			validateCrd(t, tt.wantErrors, g, clientSettingsPolicy, k8sClient)
 		})
 	}
 }
@@ -129,7 +135,16 @@ func TestClientSettingsPoliciesTargetRefGroup(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			validateClientSettingsPolicy(t, tt, g, k8sClient)
+			policySpec := tt.policySpec
+			policySpec.TargetRef.Name = gatewayv1alpha2.ObjectName(uniqueResourceName(testTargetRefName))
+			clientSettingsPolicy := &ngfAPIv1alpha1.ClientSettingsPolicy{
+				ObjectMeta: controllerruntime.ObjectMeta{
+					Name:      uniqueResourceName(testPolicyName),
+					Namespace: defaultNamespace,
+				},
+				Spec: policySpec,
+			}
+			validateCrd(t, tt.wantErrors, g, clientSettingsPolicy, k8sClient)
 		})
 	}
 }
@@ -191,46 +206,16 @@ func TestClientSettingsPoliciesKeepAliveTimeout(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			validateClientSettingsPolicy(t, tt, g, k8sClient)
+			policySpec := tt.policySpec
+			policySpec.TargetRef.Name = gatewayv1alpha2.ObjectName(uniqueResourceName(testTargetRefName))
+			clientSettingsPolicy := &ngfAPIv1alpha1.ClientSettingsPolicy{
+				ObjectMeta: controllerruntime.ObjectMeta{
+					Name:      uniqueResourceName(testPolicyName),
+					Namespace: defaultNamespace,
+				},
+				Spec: policySpec,
+			}
+			validateCrd(t, tt.wantErrors, g, clientSettingsPolicy, k8sClient)
 		})
-	}
-}
-
-func validateClientSettingsPolicy(t *testing.T, tt struct {
-	policySpec ngfAPIv1alpha1.ClientSettingsPolicySpec
-	name       string
-	wantErrors []string
-}, g *WithT, k8sClient client.Client,
-) {
-	t.Helper()
-
-	policySpec := tt.policySpec
-	policySpec.TargetRef.Name = gatewayv1alpha2.ObjectName(uniqueResourceName(testTargetRefName))
-	policyName := uniqueResourceName(testPolicyName)
-
-	clientSettingsPolicy := &ngfAPIv1alpha1.ClientSettingsPolicy{
-		ObjectMeta: controllerruntime.ObjectMeta{
-			Name:      policyName,
-			Namespace: defaultNamespace,
-		},
-		Spec: policySpec,
-	}
-	timeoutConfig := framework.DefaultTimeoutConfig()
-	ctx, cancel := context.WithTimeout(context.Background(), timeoutConfig.KubernetesClientTimeout)
-	err := k8sClient.Create(ctx, clientSettingsPolicy)
-	defer cancel()
-
-	// Clean up after test
-	defer func() {
-		_ = k8sClient.Delete(context.Background(), clientSettingsPolicy)
-	}()
-
-	if len(tt.wantErrors) == 0 {
-		g.Expect(err).ToNot(HaveOccurred())
-	} else {
-		g.Expect(err).To(HaveOccurred())
-		for _, wantError := range tt.wantErrors {
-			g.Expect(err.Error()).To(ContainSubstring(wantError), "Expected error '%s' not found in: %s", wantError, err.Error())
-		}
 	}
 }
